Apply fixes from PQClean

+++ shake256-avx2/fips202x4.c
@@ -38,7 +38,7 @@ static void keccak_absorb4x(__m256i *s,
                           const unsigned char *m1,
                           const unsigned char *m2,
                           const unsigned char *m3,
-                          unsigned long long int mlen,
+                          size_t mlen,
                           unsigned char p)
 {
   unsigned char t0[200] = {0};
@@ -147,7 +147,7 @@ void SPX_shake128x4(unsigned char *out0,
     s[i] = _mm256_xor_si256(s[i], s[i]);
 
   /* absorb 4 message of identical length in parallel */
-  keccak_absorb4x(s, SHAKE128_RATE, in0, in1, in2, in3, inlen, 0x1F);
+  keccak_absorb4x(s, SHAKE128_RATE, in0, in1, in2, in3, (size_t)inlen, 0x1F);
 
   /* Squeeze output */
   keccak_squeezeblocks4x(out0, out1, out2, out3, outlen/SHAKE128_RATE, s, SHAKE128_RATE);
@@ -194,7 +194,7 @@ void SPX_shake256x4(unsigned char *out0,
     s[i] = _mm256_xor_si256(s[i], s[i]);
 
   /* absorb 4 message of identical length in parallel */
-  keccak_absorb4x(s, SHAKE256_RATE, in0, in1, in2, in3, inlen, 0x1F);
+  keccak_absorb4x(s, SHAKE256_RATE, in0, in1, in2, in3, (size_t)inlen, 0x1F);
 
   /* Squeeze output */
   keccak_squeezeblocks4x(out0, out1, out2, out3, outlen/SHAKE256_RATE, s, SHAKE256_RATE);
